# backend/openapi.json
{
  "components": {
    "schemas": {
      "ApproveRefundRequestDto": {
        "properties": {
          "admin_notes": {
            "maxLength": 500,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "RejectRefundRequestDto": {
        "properties": {
          "admin_notes": {
            "maxLength": 500,
            "type": "string"
          }
        },
        "required": [
          "admin_notes"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "firebase": {
        "bearerFormat": "Firebase ID token",
        "scheme": "bearer",
        "type": "http"
      },
      "idempotency-key": {
        "in": "header",
        "name": "Idempotency-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "TurfX Nashik MVP — booking, discovery, partner catalog, payments, notifications.\n\nAll authenticated endpoints require a Firebase ID token via `Authorization: Bearer <token>`. See Part 3.0 §4 for the response envelope and Part 3.0 §5 for the error taxonomy.",
    "title": "TurfX API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/refunds": {
      "get": {
        "operationId": "RefundAdminController_list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/admin/refunds/{id}": {
      "get": {
        "operationId": "RefundAdminController_getOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/admin/refunds/{id}/actions/approve": {
      "post": {
        "operationId": "RefundAdminController_approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRefundRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/admin/refunds/{id}/actions/reject": {
      "post": {
        "operationId": "RefundAdminController_reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRefundRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Same origin",
      "url": "/"
    }
  ],
...
}